モックサーバーのデータをAPI定義に沿うように修正
