+++ mock/bin/moduleAlias.ts
@@ -9,4 +9,6 @@ import moduleAlias from 'module-alias'
 moduleAlias.addAliases({
   '@': path.resolve(__dirname, '../'),
   '~': path.resolve(__dirname, '../'),
+  '~openapi/*': path.resolve(__dirname, '../../openapi'),
+  '@openapi/*': path.resolve(__dirname, '../../openapi'),
 })
